requirements.txt
google-generativeai
praw
python-dotenv
tqdm
